requirements.txt
# torch
torchvision
torchinfo
h5py
seaborn
matplotlib
mlxtend
tqdm
sklearn